update on AccessLogPolicy and TargetGroupPolicy e2e-tests (#527)
